Improve socket interface
